Reorganizing Sphinx Directory for better builds and easier management
This moves all of the RST files in a _source directory which avoids a lot
of unecessary errors messages because before the make command was looking
in directories it should not have been. This also cleans up the look of
the repo and makes it easier to identify the directories that actually make
up the organization of the site.

This also adds a script that converts Google Sheets to local CSVs, though
that is still under development